refactor: change Profile -> MyProfile and PublicProfile route
